requirements.txt
blinker~=1.8.2
click~=8.1.7
colorama==0.4.6
Flask~=2.3.2
itsdangerous==2.2.0
Jinja2~=3.1.4
MarkupSafe~=3.0.2
python-dotenv~=1.0.1
Werkzeug~=3.0.0

pip~=25.1.1
attrs~=24.2.0
pillow~=10.4.0
redis~=5.1.1
typing_extensions~=4.12.2
SQLAlchemy~=2.0.25